backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
azure-ai-inference==1.0.0b6
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.7.0
azure-identity==1.19.0
azure-search-documents==11.6.0
azure-storage-blob==12.24.0
black==24.10.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coverage==7.13.5
cryptography==45.0.7
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
email_validator==2.2.0
fastapi==0.128.0
filelock==3.25.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.18
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
jiter==0.13.0
msal==1.31.0
msal-extensions==1.3.1
multidict==6.7.1
mypy==1.14.0
mypy_extensions==1.1.0
nodeenv==1.10.0
openai==2.9.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.0.1
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.2.6
pycparser==3.0
pydantic==2.12.3
pydantic-settings==2.6.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.12.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-discovery==1.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
ruff==0.8.5
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tenacity==9.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.10.8
uvicorn==0.35.0
virtualenv==21.2.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
yarl==1.23.0
